First pass at a prometheus metrics gatherer for the splunk forwarder
It's going to run inside the container and write to the host exporter's
textfile dir that's going to be mounted in the container.  Splunk doesn't
seem to provide any good way to collect metrics about the forwarder, so
to start off with, we'll collect:
    * tx/rx bytes
    * tx/rx packets
    * a splunk query of the count of messages for this host in the last hour

Since the whole purpose of this container is to send log messages to the
Splunk indexer(s), any network traffic seen in the container's network device
must be for forwarding these messages.

As for the message count.  Hopefully the query for the messages in the last
hour is a small enough impact that it won't cause a problem when deployed
across all our hosts

But for now, it's just printing the metrics it's collected to stdout to see
that it's working.

ITDEV-29896